more RF models, some lmer stuff
Mostly running more RF models
moved some files to repetition dir
r model for ns(time)*dirC

note that each RF model has tweaks to the general regionSet decoding script and instead of pushing a bunch of changes that should be viewed as alternatives rather than a series of sequential edits, i copied each script to the data folder of the relevant RF run (in ratterdam\repetition_decoding)

making changes to how the dataframe for R is set up after this push. Previously we assign a field to 1 alley and if it overlaps above thresh >1 alley we ignore. however this is throwing out too much data and more importantly it seems like fields wrapping around corners, extending into 2 alleys, is a thing we want to look at. so we will just loop over alleys cell is in and sort it out in R by separating by dirC. this will be done in new script.