starting winding prototyping, pushing before refactoring alleybounds code
Not a ton new. Starting to analyze winding data, some changes to preprocessing loading. Major reason for pushing is checkpoint because about to refactor the alleybounds code. Recently have been loading one set of bounds for each rat but that doesnt work if different days need different bounds (which is the case for R886).