diagnostic plots
worked on diagnostic plots of field traversals and behavior over time (separate files). think the field traversal code + viz is finally working (with possible edge case of some n+1 turn turnarounds not being handled right?).

Also there was a bug in the regionSet decoding that used a filtered turn df but thats no good bc the filtered df is not continuous so n+1 in df is not the next turn necessarily. fixed here.