ethogram stuff and prepping to refactor RF decoding
started looking at neural activity by frequency of behavior (although a lot of this was in the console).

going to refactor / functionalize decoder (RF and naive) to make pipeline easier to run bc I am about to run a bunch of combinations of things. so pushing before i break stuff.