Finalizing preprocessing for repetition manuscript
Tweaking and checking preprocessing scripts to finalize the data creation. Mostly related to creating the "superpop" csv. This is a table of all rats and days. Every entry is a pass through a field labeled with the normalized firing rate (normed by location thru field) and attributes like route headings, time, reward, etc

There were some things to fix. Like a camera offset for R886D1 was fixed twice, which offset it in the opposite direction.

Creating file for time dynamics figure. In this I fixed the blowup problem with pchip interpolation.

william (snider) defaults "created" here, it was previously in another directory.