Rerunning analyses for paper, modifications to analyses
Rerunning analyses for paper. Mostly changes to preprocessing, the biggest being I was previously removing lost tracking (turn.Ego==3) before creating turn df, which introduced discontinuities. Fixed that.

Small tweaks to analysis. Changed bootstrapping (which was really a hybrid bootstrapping + shuffling) to be more just a shuffling analysis. File and variable names unchanged yet, need to port it to a new script or something.

Redid random forest to use trials (passes thru alley) normalized by overall firing across the session in that alley. So equating it with the superpop.csv we use for GLMs and MW.

Did shuffling of GLM emmeans analysis of CD, PD, ND. Seems like no effect of ND and almost an effect of PD. Still an effect of CD.