Working on paper figures
Making paper figures and doing last analyses of repetition.
Changes to existing files are adding more to figure panels, making analyses better
new scripts are for new analyses like the totalpaths/unique paths stuff

may rework (after this commit) the figure code scripts a fair amount as the later figures in particular have gotten messy and not all of it is even in the right files (the files in the repetition_manuscript folder).